move middleware, actions, and api reducer into api-state directory
