+++ src/api-state/cache/index.test.js
@@ -7,12 +7,11 @@ import {
 	MOCK_API_RESULT,
 } from 'meetup-web-mocks/lib/app';
 
-import { epicIgnoreAction } from '../util/testUtils';
+import { epicIgnoreAction } from '../../util/testUtils';
+import * as api from '../sync/apiActionCreators';
 
-import { makeCache } from '../util/cacheUtils';
-
-import getCacheEpic from './cache';
-import * as api from '../actions/apiActionCreators';
+import { makeCache } from './util';
+import getCacheEpic from './';
 
 const MOCK_QUERY = mockQuery(MOCK_RENDERPROPS);
 const MOCK_SUCCESS_ACTION = api.success({
