+++ src/api-state/index.js
@@ -1,8 +1,28 @@
 import { combineEpics, createEpicMiddleware } from 'redux-observable';
 
-import getSyncEpic from '../epics/sync';
-import getCacheEpic from '../epics/cache';
-import { postEpic, deleteEpic } from '../epics/mutate'; // DEPRECATED
+import getSyncEpic from './sync';
+import getCacheEpic from './cache';
+import { postEpic, deleteEpic } from './mutate'; // DEPRECATED
+
+// export specific values of internal modules
+export {
+	LOCATION_CHANGE,
+	SERVER_RENDER,
+	locationChange,
+} from './sync/syncActionCreators';
+export {
+	API_REQ,
+	API_RESP_SUCCESS,
+	API_RESP_COMPLETE,
+	API_RESP_ERROR,
+	API_RESP_FAIL,
+	requestAll,
+	get,
+	post,
+	patch,
+	del,
+} from './sync/apiActionCreators';
+export { api, app } from './reducer';
 
 /**
  * The middleware is exported as a getter because it needs the application's
@@ -13,7 +33,7 @@ import { postEpic, deleteEpic } from '../epics/mutate'; // DEPRECATED
  * order to render the application. We may want to write a server-specific
  * middleware that doesn't include the other epics if performance is an issue
  */
-const getPlatformMiddleware = (routes, fetchQueries, baseUrl) =>
+export const getApiMiddleware = (routes, fetchQueries, baseUrl) =>
 	createEpicMiddleware(
 		combineEpics(
 			getSyncEpic(routes, fetchQueries, baseUrl),
@@ -22,5 +42,3 @@ const getPlatformMiddleware = (routes, fetchQueries, baseUrl) =>
 			deleteEpic // DEPRECATED
 		)
 	);
-
-export default getPlatformMiddleware;
