+++ src/util/createStoreServer.js
@@ -1,6 +1,7 @@
 import { applyMiddleware, createStore } from 'redux';
+import { getApiMiddleware } from '../api-state'; // mwp-api-state
+
 import { parseQueryResponse } from './fetchUtils';
-import getEpicMiddleware from '../middleware/epic';
 import catchMiddleware from '../middleware/catch';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/toPromise';
@@ -32,7 +33,7 @@ export const serverFetchQueries = request => () => queries =>
 export function getServerCreateStore(routes, middleware, request, baseUrl) {
 	const middlewareToApply = [
 		catchMiddleware(err => request.server.app.logger.error(err)),
-		getEpicMiddleware(routes, serverFetchQueries(request), baseUrl),
+		getApiMiddleware(routes, serverFetchQueries(request), baseUrl),
 		...middleware,
 	];
