+++ src/util/fetchUtils.js
@@ -10,29 +10,6 @@ import { setClickCookie } from '../plugins/tracking/util/clickState';
 export const CSRF_HEADER = 'x-csrf-jwt';
 export const CSRF_HEADER_COOKIE = 'x-csrf-jwt-header';
 
-/**
- * @deprecated
- */
-export function getDeprecatedSuccessPayload(successes, errors) {
-	const allQueryResponses = [...successes, ...errors];
-	return allQueryResponses.reduce(
-		(payload, { query, response }) => {
-			if (!response) {
-				return payload;
-			}
-			const { ref, error, ...responseBody } = response;
-			if (error) {
-				// old payload expects error as a property of `value`
-				responseBody.value = { error };
-			}
-			payload.queries.push(query);
-			payload.responses.push({ [ref]: responseBody });
-			return payload;
-		},
-		{ queries: [], responses: [] }
-	);
-}
-
 export const parseQueryResponse = queries => ({
 	responses,
 	error,
